Don't crash for missing config things

+++ downmail/filter.py
@@ -17,35 +17,36 @@ class Filter(object):
 
                 # Filter out blacklisted senders and phrases
                 for message in itertools.islice(dm_account.get_unanswered_messages(), 100):
-                    if message.sender_address in filter_group['simple_blacklist']:
+                    if 'simple_blacklist' in filter_group and message.sender_address in filter_group['simple_blacklist']:
                         message.delete()
                         continue
 
-                    for blacklist_rule in filter_group['blacklist']:
-                        should_delete = False
-                        if 'to' in blacklist_rule:
-                            if message.recipient == blacklist_rule['to']:
-                                should_delete = True
-
-                        if 'from' in blacklist_rule:
-                            if message.sender_address == blacklist_rule['from']:
-                                should_delete = True
-
-                        if should_delete and 'contains' in blacklist_rule:
+                    if 'blacklist' in filter_group:
+                        for blacklist_rule in filter_group['blacklist']:
                             should_delete = False
-                            cant_contain = blacklist_rule['contains']
-                            for badphrase in cant_contain:
-                                if (message.subject + message.text).lower().count(badphrase):
+                            if 'to' in blacklist_rule:
+                                if message.recipient == blacklist_rule['to']:
                                     should_delete = True
 
-                        if should_delete:
-                            print(message)
-                            message.delete()
+                            if 'from' in blacklist_rule:
+                                if message.sender_address == blacklist_rule['from']:
+                                    should_delete = True
 
+                            if should_delete and 'contains' in blacklist_rule:
+                                should_delete = False
+                                cant_contain = blacklist_rule['contains']
+                                for badphrase in cant_contain:
+                                    if (message.subject + message.text).lower().count(badphrase):
+                                        should_delete = True
+
+                            if should_delete:
+                                print(message)
+                                message.delete()
 
                     # TODO apply replacements by copying the message with all of its payloads, replacing all instances of the problem with {What you want instead}, and sending that to self
-                    for replacement_rule in filter_group['replacements']:
-                        pass
+                    if 'replacements' in filter_group:
+                        for replacement_rule in filter_group['replacements']:
+                            pass
